requirements.txt
psutil>=5.9.8
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
grpcio>=1.40.0
grpcio-tools>=1.40.0
pydantic>=1.8.0
requests>=2.25.0
httpx>=0.24.0
pytest-asyncio>=0.21
pyOpenSSL>=21.0.0
etcd3gw>=2.4.0
watchdog>=3.0.0
